+++ changelog.md
@@ -21,3 +21,4 @@
 5. `Scene.set_object_orientation` and `Scene.set_object_local_scaling` supports tuple of floats as well as inputs, additional to `jnp.array`.
 6. `Model` now has a convenient method `create` to create a Model with same face indices shared by `faces`, `faces_norm` and `faces_uv`, and a default `specular_map`. This is useful for creating a mesh where all vertices has its own normal and uv coordinate specified, under same order (thus same face indices).
 7. Correctly support Python Sequence for `utils.build_texture_from_PyTinyrenderer` as texture.
+8. `quaternion` function to create an orientation from axis and angle.